feat: done with the queue system, probably have some stuff to cleanup later but for now this is nice
