+++ package.json
@@ -38,7 +38,6 @@
     "redis": "^4.3.1",
     "string-similarity": "^4.0.4",
     "subtitle": "^4.2.1",
-    "threads": "^1.7.0",
     "uuid": "^8.3.2",
     "vue": "^3.2.37",
     "vue-i18n": "^9.1.10",
