+++ worker-layer/index.ts
@@ -1,6 +1,8 @@
 import "dotenv/config";
 import "#/workers/processSources";
 import "#/workers/scanDirectoryForEntities";
+import "#/workers/convertFileToWebMp4";
+import "#/workers/bindFileToDbEntity";
 import { QueueProcessSources } from "./queues/processSources";
 
 import { addTvDbTokenToProcessEnv } from "~/addTvDbTokenToProcessEnv";
@@ -8,15 +10,12 @@ import { addTvDbTokenToProcessEnv } from "~/addTvDbTokenToProcessEnv";
 addTvDbTokenToProcessEnv().then(() => {
   const queue = QueueProcessSources();
 
-  console.log("additon shit to q");
   // Initial job calls
   queue.add("processSources", undefined, {
     // Repeat scan sources every 5 hours
     repeat: {
       every: 1000 * 60 * 60 * 5,
-      limit: 2,
+      immediately: true,
     },
   });
-  // Imediate queue
-  queue.add("processSources", undefined);
 });
