+++ worker-layer/processEntityHelper/fromBestMatchFillDb.ts
@@ -1,25 +1,34 @@
 import { GetRouteDataHandlerFromUrlAndVerb } from "~/types/Route";
 import { AllRoutes } from "~/types/RouteLibraryDataLayer";
 
-import { processIdAsMovie } from "./movie";
-import { processIdAsEpisode } from "./episode";
+import { processIdAsMovie } from "#/processEntityHelper/processIdAsMovie";
+import { processIdAsEpisode } from "#/processEntityHelper/processIdAsEpisode";
+import { PrismaClient, Episode, Movie, SerieV2 } from "@prisma/client";
 
-import { userIsAdmin } from "~/middlewares/userIsAdmin";
+type ProcessEntityPayload<T extends "movie" | "serie" = "movie" | "serie"> = {
+  entityId: `${T}-${number}`;
+  episodeInfo: T extends "movie" ? undefined : EpisodeIndexInfo;
+};
 
-export const processEntityIdPost: GetRouteDataHandlerFromUrlAndVerb<
-  "post",
-  AllRoutes,
-  "/processEntity"
-> = async (prisma, req, res, payload) => {
-  if (!(await userIsAdmin(prisma, req)))
-    return res.status(403).json("Not allowed") as any;
-  const { entityId, episodeInfo } = payload;
+export type EpisodeIndexInfo = {
+  episodeNumber: number;
+  seasonNumber: number;
+};
+
+const prisma = new PrismaClient();
+
+export async function fromBestMatchFillDb<T extends "movie" | "serie">({
+  entityId,
+  episodeInfo,
+}: ProcessEntityPayload<T>): Promise<T extends "serie" ? Episode : Movie> {
   const { type, id } = getTypeAndIdFromEntityId(entityId);
 
+  // @ts-expect-error
   if (type === "movie") return await processIdAsMovie(prisma, id);
   if (!episodeInfo) throw new Error("Missing episode infos");
+  // @ts-expect-error
   return await processIdAsEpisode(prisma, id, episodeInfo);
-};
+}
 
 function getTypeAndIdFromEntityId(entityId: string | undefined): {
   type: "movie" | "serie";
