+++ worker-layer/transactions/upsertAndConnectSerieOverviewTranslationCollection.ts
