+++ worker-layer/tvdb-api/getTvDbMovieFomId.ts
@@ -1,4 +1,4 @@
-import { tvDbGetRequest } from "#/tvDbApi";
+import { tvDbGetRequest } from "~/tvDbApi";
 
 import { TvDbMovie } from "~/types/RouteLibraryTvDbApi";
