+++ worker-layer/workers/convertFileToWebMp4.ts
@@ -1,25 +1,22 @@
 import path from "node:path";
-import { expose } from "threads";
 import { spawn, execSync } from "node:child_process";
+import { Worker } from "bullmq";
+import { redisConfig } from "#/redisConfig";
 
-export type ConvertFileToMp4ThreadWorkerReturn = void;
-export type ConvertFileToMp4ThreadWorker = (
-  filePath: string
-) => Promise<ConvertFileToMp4ThreadWorkerReturn>;
+import { QueueBindFileToEntity } from "#/queues/bindFileToDbEntity";
+import { findMatchingSubTracks } from "#/processEntityHelper/findMatchingSubTracksFromFullPath";
 
-enum MachineHardwareType {
-  Nvidea,
-  Amd,
-  Cpu,
-}
+export type ConvertFileToWebMp4Return = void;
+export type ConvertFileToWebMp4Data = { filePath: string; sourcePath: string };
+export type ConvertFileToWebMp4QueueName = "convertFileToWebMp4";
 
-const convertFileToMp4ThreadWorker: ConvertFileToMp4ThreadWorker = async (
-  filePath
-) => {
+const convertFileToWebMp4: (
+  data: ConvertFileToWebMp4Data
+) => Promise<ConvertFileToWebMp4Return> = async ({ filePath, sourcePath }) => {
   const parsedPath = path.parse(filePath);
   const newFileName = `${parsedPath.dir}/${parsedPath.name}.mp4`;
   const removeCommand = `rm "${filePath}"`;
-  const command = await getFfmpegCommand(filePath, newFileName);
+  const command = getFfmpegCommand(filePath, newFileName);
   let fileIsMostProbablyGoodToGo = false;
 
   return new Promise((resolve, reject) => {
@@ -27,15 +24,13 @@ const convertFileToMp4ThreadWorker: ConvertFileToMp4ThreadWorker = async (
     let NUMBER_OF_FRAMES: number | undefined;
 
     console.log(`Started conversion for ${path.basename(filePath)}`);
-    console.log({ command });
     const child = spawn(command, {
       shell: true,
     });
 
-    child.stderr.on("data", (err: Buffer) => {
-      const message = err.toString("utf-8");
+    child.stderr.on("data", (buffer: Buffer) => {
+      const message = buffer.toString("utf-8");
 
-      // console.log({ message });
       if (NUMBER_OF_FRAMES === undefined) {
         if (!message.includes("NUMBER_OF_FRAMES")) return;
         const regex = /NUMBER_OF_FRAMES(-\D*)?:(\s*)?(?<frame>\d*)/gi;
@@ -45,8 +40,6 @@ const convertFileToMp4ThreadWorker: ConvertFileToMp4ThreadWorker = async (
         NUMBER_OF_FRAMES = Number(matches.groups.frame);
       } else {
         if (!message.includes("frame=")) return;
-        // At this point it's pretty sure that the conversion is going well, so we can flag the source file as good to be deleted
-        fileIsMostProbablyGoodToGo = true;
         const regex = /frame=(\s*)?(?<frame>\d*)/gi;
         const matches = regex.exec(message);
 
@@ -66,26 +59,45 @@ const convertFileToMp4ThreadWorker: ConvertFileToMp4ThreadWorker = async (
     });
 
     child.on("close", () => {
+      console.log("close called");
       resolve();
     });
 
-    child.on("exit", () => {
-      if (fileIsMostProbablyGoodToGo) execSync(removeCommand);
+    child.on("exit", (returnCode) => {
+      if (returnCode === 0) {
+        // execSync(removeCommand);
+        const registerEntityQueue = QueueBindFileToEntity();
+
+        registerEntityQueue.add("bindFileToDbEntity", {
+          filePath: newFileName,
+          potencialSubtitleTrack: findMatchingSubTracks(newFileName),
+          sourcePath,
+        });
+      }
       resolve();
     });
 
     child.on("error", (err) => {
-      console.log("error called");
+      console.log("error called", err);
       reject(err);
     });
   });
 };
 
-async function getFfmpegCommand(
-  filePath: string,
-  newFileName: string
-): Promise<string> {
+function getFfmpegCommand(filePath: string, newFileName: string): string {
   return `ffmpeg -vaapi_device /dev/dri/renderD128 -i "${filePath}" -vf 'format=nv12,hwupload' -c:v h264_vaapi "${newFileName}" -y`;
 }
 
-expose(convertFileToMp4ThreadWorker);
+new Worker<
+  ConvertFileToWebMp4Data,
+  ConvertFileToWebMp4Return,
+  ConvertFileToWebMp4QueueName
+>(
+  "convertFileToWebMp4",
+  async ({ data }) => {
+    console.log(`Adding "${data.filePath}" to get converted`);
+    await convertFileToWebMp4(data);
+    console.log(`"${data.filePath}" conversion job finished`);
+  },
+  { ...redisConfig, concurrency: 4 }
+);
